--- backend/controllers/authController.js
@@ -1,79 +1,0 @@
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const User = require("../models/userModel");
-const createError = require("../utils/appError");
-
-//Register user
-exports.signup = async (req, res, next) => {
-  try {
-    const { name, email, password, confirmPassword } = req.body;
-
-    // Check if passwords match
-    if (password !== confirmPassword) {
-      return next(new createError("Passwords do not match", 400));
-    }
-
-    // Check if user already exists
-    const existingUser = await User.findOne({ email });
-    if (existingUser) {
-      return next(new createError("User already exists", 400));
-    }
-
-    // Hash password
-    const hashedPassword = await bcrypt.hash(password, 12);
-
-    // Create new user
-    const newUser = await User.create({
-      name,
-      email,
-      password: hashedPassword,
-    });
-
-    // Assign JWT token
-    const token = jwt.sign({ id: newUser._id }, process.env.JWT_SECRET, {
-      expiresIn: process.env.JWT_EXPIRES_IN,
-    });
-
-    // Send response
-    res.status(201).json({
-      status: "success",
-      message: "User registered successfully",
-      token,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-// Login user
-exports.signin = async (req, res, next) => {
-  try {
-    const { email, password } = req.body;
-
-    // 1. Check if user exists
-    const user = await User.findOne({ email });
-    if (!user) {
-      return next(new createError("User not found", 404));
-    }
-
-    // 2. Check password
-    const isMatch = await bcrypt.compare(password, user.password);
-    if (!isMatch) {
-      return next(new createError("Invalid email or password", 401));
-    }
-
-    // 3. Generate JWT
-    const token = jwt.sign({ id: user._id }, process.env.JWT_SECRET, {
-      expiresIn: process.env.JWT_EXPIRES_IN,
-    });
-
-    // 4. Send response
-    res.status(200).json({
-      status: "success",
-      message: "Logged in successfully",
-      token,
-    });
-  } catch (error) {
-    next(error);
-  }
-};
