--- backend/models/userModel.js
@@ -1,19 +1,0 @@
-const mongoose = require("mongoose");
-
-const userSchema = new mongoose.Schema({
-    name:{
-        type: String,
-        required: true,
-    },
-    email:{
-        type: String,
-        unique: true,
-        required: true,
-    },
-    password:{
-        type: String,
-        required: true,
-    },
-})
-const User = mongoose.model('User',userSchema);
-module.exports = User;
