--- backend/utils/appError.js
@@ -1,9 +1,0 @@
-class createError extends Error {
-  constructor(message, statusCode) {
-    super(message);
-    this.statusCode = statusCode;
-    this.status = `${statusCode}`.startsWith("4") ? "fail" : "error";
-    Error.captureStackTrace(this, this.constructor);
-  }
-}
-module.exports = createError;
