+++ frontend/index.html
@@ -48,7 +48,7 @@
             <a href="./pages/feedback.html" class="navbar__links">Feedback</a>
           </li>
           <li class="navbar__btn">
-            <a href="./pages/signin.html" class="button">SignIn</a>
+            <a href="./pages/sign.html" class="button">SignIn</a>
           </li>
         </ul>
       </div>
