--- frontend/pages/signin.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <link rel="stylesheet" href="../styles/sign.css" />
-    <title>Sign In</title>
-  </head>
-  <body>
-    <div class="login-wrap">
-      <div class="login-html">
-        <label class="tab">Sign In</label>
-        <div class="login-form">
-          <form id="signin-form" class="sign-in-htm">
-            <div class="group">
-              <label for="email" class="label">Email</label>
-              <input
-                id="email"
-                name="email"
-                type="email"
-                class="input"
-                required
-              />
-            </div>
-            <div class="group">
-              <label for="password" class="label">Password</label>
-              <input
-                id="password"
-                name="password"
-                type="password"
-                class="input"
-                required
-              />
-            </div>
-            <div class="group">
-              <input type="submit" class="button" value="Sign In" />
-            </div>
-          </form>
-        </div>
-      </div>
-    </div>
-  </body>
-  <script>
-    document
-      .getElementById("signin-form")
-      .addEventListener("submit", async (e) => {
-        e.preventDefault();
-
-        const email = document.getElementById("email").value;
-        const password = document.getElementById("password").value;
-
-        try {
-          const res = await fetch("/api/auth/signin", {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ email, password }),
-          });
-
-          const data = await res.json();
-          if (res.ok) {
-            alert("Login successful!");
-            window.location.href = "/index.html";
-          } else {
-            alert(data.message || "Login failed");
-          }
-        } catch (err) {
-          console.error(err);
-          alert("Something went wrong");
-        }
-      });
-  </script>
-</html>
