--- frontend/pages/signup.html
@@ -1,95 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <link rel="stylesheet" href="../styles/sign.css" />
-    <title>Sign Up</title>
-  </head>
-  <body>
-    <div class="login-wrap">
-      <div class="login-html">
-        <label class="tab">Sign Up</label>
-        <div class="login-form">
-          <form id="signup-form" class="sign-up-htm">
-            <div class="group">
-              <label for="user" class="label">Username</label>
-              <input id="user" name="name" type="text" class="input" required />
-            </div>
-            <div class="group">
-              <label for="pass" class="label">Password</label>
-              <input
-                id="pass"
-                name="password"
-                type="password"
-                class="input"
-                required
-              />
-            </div>
-            <div class="group">
-              <label for="repeat-pass" class="label">Repeat Password</label>
-              <input
-                id="repeat-pass"
-                name="confirmPassword"
-                type="password"
-                class="input"
-                required
-              />
-            </div>
-            <div class="group">
-              <label for="email" class="label">Email Address</label>
-              <input
-                id="email"
-                name="email"
-                type="email"
-                class="input"
-                required
-              />
-            </div>
-            <div class="group">
-              <input type="submit" class="button" value="Sign Up" />
-            </div>
-            <div class="hr"></div>
-            <div class="foot-lnk">
-              <a href="./sign-in.html">Already a member? Sign In</a>
-            </div>
-          </form>
-        </div>
-      </div>
-    </div>
-  </body>
-  <script>
-    document
-      .getElementById("signup-form")
-      .addEventListener("submit", async (e) => {
-        e.preventDefault();
-
-        const name = document.getElementById("user").value;
-        const email = document.getElementById("email").value;
-        const password = document.getElementById("pass").value;
-        const confirmPassword = document.getElementById("repeat-pass").value;
-
-        if (password !== confirmPassword) {
-          alert("Passwords do not match!");
-          return;
-        }
-
-        try {
-          const res = await fetch("/api/auth/signup", {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ name, email, password }),
-          });
-
-          const data = await res.json();
-          if (res.ok) {
-            alert("Sign-up successful!");
-            window.location.href = "/index.html";
-          } else {
-            alert(data.message || "Sign-up failed");
-          }
-        } catch (err) {
-          console.error(err);
-          alert("Something went wrong.");
-        }
-      });
-  </script>
-</html>
